requirements.txt
streamlit
pandas
numpy
folium
streamlit-folium
requests
matplotlib
scipy
python-dotenv